+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add DiscV5 discovery metrics (`discv5_live_nodes_current`, `discv5_total_nodes_current`) to track node counts in the routing table [#9692](https://github.com/besu-eth/besu/issues/9692)
 - Add `txpool_contentFrom` JSON-RPC method [#10111](https://github.com/besu-eth/besu/pull/10111)
 - Add `txpool_content` JSON-RPC method [#10120](https://github.com/besu-eth/besu/pull/10120)
+- Add `txpool_inspect` JSON-RPC method [#10121](https://github.com/besu-eth/besu/pull/10121)
 
 ## 26.3.0